refactor(matcha-window): split `winit_interface` into common/native/web
Moves the platform-divergent entry point (`run`, which returns
`Result<_, EventLoopError>` on native and `()` on wasm via `spawn_app`)
into `winit_interface/native.rs` and `winit_interface/web.rs`. The
remaining ~470 lines (`WinitInterface` struct, `ApplicationHandler`
impl, `EventLoop`/`EventLoopProxy` impls for winit types, event
mapping helpers, `ControlFlow` conversions) move to
`winit_interface/common.rs` and are now cfg-free.
